+++ pkg/aggregator/rate.go
@@ -7,19 +7,20 @@ import (
 // Rate tracks the rate of a metric over 2 successive flushes
 type Rate struct {
 	previousSample    float64
-	previousTimestamp int64
+	previousTimestamp float64
 	sample            float64
-	timestamp         int64
+	timestamp         float64
 }
 
-func (r *Rate) addSample(sample *MetricSample, timestamp int64) {
+func (r *Rate) addSample(sample *MetricSample, timestamp float64) {
 	if r.timestamp != 0 {
 		r.previousSample, r.previousTimestamp = r.sample, r.timestamp
 	}
 	r.sample, r.timestamp = sample.Value, timestamp
+
 }
 
-func (r *Rate) flush(timestamp int64) ([]*Serie, error) {
+func (r *Rate) flush(timestamp float64) ([]*Serie, error) {
 	if r.previousTimestamp == 0 || r.timestamp == 0 {
 		return []*Serie{}, NoSerieError{}
 	}
@@ -28,9 +29,9 @@ func (r *Rate) flush(timestamp int64) ([]*Serie, error) {
 		return []*Serie{}, fmt.Errorf("Rate was sampled twice at the same timestamp, can't compute a rate")
 	}
 
-	value, ts := (r.sample-r.previousSample)/float64(r.timestamp-r.previousTimestamp), r.timestamp
+	value, ts := (r.sample-r.previousSample)/(r.timestamp-r.previousTimestamp), r.timestamp
 	r.previousSample, r.previousTimestamp = r.sample, r.timestamp
-	r.sample, r.timestamp = 0, 0
+	r.sample, r.timestamp = 0., 0.
 
 	return []*Serie{
 		&Serie{
